Fix parsing of "= version =" headers in readme.txt
The readme.txt format that's used by WordPress is not quite standard Markdown. It includes a couple of features that are not directly supported by popular Markdown parsers. One of those features is parsing "= string =" as "<h4>string</h4>" (the opening "=" must be at the start of the line). Leading whitespace is allowed. This syntax is typically for plugin version numbers in the changelog section.

The problem: the readme parser discards leading whitespace when parsing these headers. This means it can also discard the previous line break character(s). As a result, this markup:
 
 * abc
 
 = 1.2.3 =
 * def
 
 Becomes this:
 
 * abc
 <h4>1.2.3</h4>
 * def
 
 The way that Parsedown handles this markup is to make the H4 a part of the preceding list item. And that's a bug. 
   
 Fixed by keeping leading whitespace instead of throwing it away.

Closes #68