VCS integration: Support alternative readme.txt capitalisations
Apparently, some developers call the file "README.txt" instead of "readme.txt". Most version control systems are case-sensitive, so we need to known the exact capitalisation to successfully retrieve the readme from the repository. Let's search the plugin directory to find the local readme file and then use that name. 

Closes #147