Fixed a potential fatal error when the current PUC version is removed during an upgrade.
By default, PUC automatically checks for updates immediately after the user upgrades the plugin or theme. This can become a problem if the upgrade overwrites the PUC library with a different version. While checking for updates, PUC may attempt to autoload a class, which will trigger a fatal error if the class file was deleted during the upgrade.

Fixed by checking if the running PUC version still exists after the upgrade. If it doesn't, remove hooks and stop.